+++ crates/blockchain/blockchain.rs
@@ -7,7 +7,7 @@ mod smoke_test;
 pub mod tracing;
 pub mod vm;
 
-use ::tracing::info;
+use ::tracing::{error, info};
 use constants::MAX_INITCODE_SIZE;
 use error::MempoolError;
 use error::{ChainError, InvalidBlockError};
@@ -143,6 +143,14 @@ impl Blockchain {
 
         query_plan.apply_to_store(self.storage.clone()).await?;
 
+        if let Err(error) = self
+            .storage
+            .add_block_snapshot(block.clone(), account_updates.to_vec())
+            .await
+        {
+            error!("Error adding block snapshot: {}", error);
+        }
+
         Ok(())
     }
 
@@ -169,7 +177,17 @@ impl Blockchain {
         self.storage
             .add_receipts(block.hash(), execution_result.receipts)
             .await
-            .map_err(ChainError::StoreError)
+            .map_err(ChainError::StoreError)?;
+
+        if let Err(error) = self
+            .storage
+            .add_block_snapshot(block.clone(), account_updates.to_vec())
+            .await
+        {
+            error!("Error adding block snapshot: {}", error);
+        }
+
+        Ok(())
     }
 
     pub async fn add_block(&self, block: &Block) -> Result<(), ChainError> {