+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Perf
 
+### 2025-05-29
+
+- Add block snapshots with diff layers. [2664](https://github.com/lambdaclass/ethrex/pull/2664)
+- Make disk layer in snapshots use the database. [2848](https://github.com/lambdaclass/ethrex/pull/2848)
+
 ### 2025-05-27
 
 - Improved the performance of shift instructions. [2933](https://github.com/lambdaclass/ethrex/pull/2933)